Liten refactoring i forbindelse med BucService
