+++ src/main/kotlin/no/nav/eessi/pensjon/klienter/eux/EuxKlient.kt
@@ -1,7 +1,11 @@
 package no.nav.eessi.pensjon.klienter.eux
 
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry
+import no.nav.eessi.pensjon.json.mapJsonToAny
+import no.nav.eessi.pensjon.json.typeRefs
 import no.nav.eessi.pensjon.metrics.MetricsHelper
+import no.nav.eessi.pensjon.service.buc.Participant
+import no.nav.eessi.pensjon.service.buc.ParticipantHolder
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -88,16 +92,18 @@ class EuxKlient(
 
     @Retryable(include = [HttpStatusCodeException::class]
             , backoff = Backoff(delay = 30000L, maxDelay = 3600000L, multiplier = 3.0))
-    fun hentBuc(bucId: String): String? {
+    fun hentInstitusjonerIBuc(bucId: String): List<Participant> {
         return hentBuc.measure {
             return@measure try {
                 logger.info("Henter buc for rinaSakId: $bucId")
 
-                euxOidcRestTemplate.exchange(
+                val buc = euxOidcRestTemplate.exchange(
                         "/buc/$bucId",
                         HttpMethod.GET,
                         null,
                         String::class.java).body
+
+                mapJsonToAny(buc!!, typeRefs<ParticipantHolder>()).participants
 
             } catch(ex: Exception) {
                 logger.warn("En feil oppstod under henting av BUC", ex)
